Replace BigInteger with common BigInt implementation

+++ src/jvmMain/kotlin/AbstractFrameBufferSurface.kt
@@ -1,7 +1,6 @@
 import TextureQualityOptionState.Companion.method1783
 import java.awt.Canvas
 import java.awt.Graphics
-import java.math.BigInteger
 
 abstract class AbstractFrameBufferSurface : LinkedListNode() {
     @JvmField
@@ -25,7 +24,7 @@ abstract class AbstractFrameBufferSurface : LinkedListNode() {
         @JvmField
         var anInt6919: Int = 0
         @JvmField
-        var aBigInteger6921: BigInteger?
+        var aBigInteger6921: BigInt?
         @JvmField
         var anInt6922: Int = 0
         @JvmField
